Big FFT dimensions support - up to pow(2,30)
-Changed the way VkFFT achieves coalesced memory accesses - FFT grouping
-Removed all transposition steps - data order is now preserved
-Added support for big FFT sequences done in multiple uploads - 5-20% speed advantage over cuFFT on 1660 Ti
-Increased Intel performance by 30% on some FFT sizes
-Big FFT C2C benchmark (sample id 4 and cuFFT script)

+++ Vulkan_FFT.cpp
@@ -987,7 +987,7 @@ int main()
 		//4 - FFT + iFFT C2C benchmark for big systems
 		const uint32_t num_benchmark_samples = 9;
 		const uint32_t num_runs = 5;
-		uint32_t benchmark_dimensions[num_benchmark_samples][4] = { {1024, 1024, 1, 2}, {pow(2,13), 32, 1, 2}, {pow(2,14), 32, 1, 2}, {pow(2,15), 32, 1, 2}, {pow(2,16), 32, 1, 2}, {pow(2,17), 32, 1, 2}, {pow(2,18), 32, 1, 2}, {pow(2,13), pow(2,13), 1, 2},{pow(2,14), pow(2,14), 1, 2} };
+		uint32_t benchmark_dimensions[num_benchmark_samples][4] = { {1024, 1024, 1, 2}, {(uint32_t) pow(2,13), 32, 1, 2}, {(uint32_t) pow(2,14), 32, 1, 2}, {(uint32_t) pow(2,15), 32, 1, 2}, {(uint32_t) pow(2,16), 32, 1, 2}, {(uint32_t) pow(2,17), 32, 1, 2}, {(uint32_t) pow(2,18), 32, 1, 2}, {(uint32_t) pow(2,13), (uint32_t) pow(2,13), 1, 2},{(uint32_t) pow(2,14), (uint32_t) pow(2,14), 1, 2} };
 																	
 
 		for (uint32_t n = 0; n < num_benchmark_samples; n++) {
